chore: moves duration out of folder
